requirements.txt
Fabric==1.10.2
boto==2.38.0
django-dotenv==1.3.0
ipython==4.0.0
Jinja2==2.8
